services/mcp-backend/requirements.txt
fastmcp
asyncpg
python-dotenv
